requirements.txt
docker==6.1.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0  # JWT支持
python-multipart==0.0.6  # 用于表单处理 